pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "grok-research-agent"
version = "0.1.0"
description = "Grok-powered research workflow automation CLI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "openai>=1.30.0",
  "python-dotenv",
  "rich",
  "pydantic>=2",
  "tiktoken",
  "pypdf",
  "requests",
  "beautifulsoup4",
  "readability-lxml",
  "chardet<6",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]


[project.scripts]
grok-research-agent = "grok_research_agent.cli:main"

[tool.setuptools]
packages = ["grok_research_agent"]
include-package-data = true

[tool.setuptools.package-data]
grok_research_agent = ["prompts/*.txt"]
